+++ src/main/java/com/eclipsekingdom/playerplot/loot/PlotDeed.java
@@ -1,7 +1,7 @@
 package com.eclipsekingdom.playerplot.loot;
 
+import com.eclipsekingdom.playerplot.sys.Language;
 import com.eclipsekingdom.playerplot.sys.Version;
-import com.eclipsekingdom.playerplot.util.ItemUtil;
 import org.bukkit.ChatColor;
 import org.bukkit.Material;
 import org.bukkit.enchantments.Enchantment;
@@ -12,41 +12,56 @@ import org.bukkit.inventory.meta.ItemMeta;
 import java.util.ArrayList;
 import java.util.List;
 
-import static com.eclipsekingdom.playerplot.sys.Language.*;
-
 public class PlotDeed {
 
-    private static String plotDeedLoreID = ChatColor.DARK_GRAY + MISC_PLOT_DEED_LORE.toString();
+    private static String plotDeedLoreID = ChatColor.DARK_GRAY + Language.MISC_PLOT_DEED_LORE.toString();
 
-    private static ItemStack itemStack = buildItemStack();
+    private ItemStack itemStack;
 
-    private static ItemStack buildItemStack() {
+    public PlotDeed(int customModelData, String title) {
+        this.itemStack = buildItemStack(customModelData, title);
+    }
+
+    private ItemStack buildItemStack(int customModelData, String title) {
         ItemStack plotDeed = new ItemStack(Material.PAPER);
         ItemMeta meta = plotDeed.getItemMeta();
-        meta.setDisplayName(ChatColor.DARK_PURPLE + LABEL_PLOT_DEED.toString());
+        meta.setDisplayName(title);
         List<String> lore = new ArrayList<>();
         lore.add(plotDeedLoreID);
-        lore.add(ChatColor.RED + MISC_ONE_USE.toString());
-        lore.add(ChatColor.GRAY + "+1 plot");
+        lore.add(ChatColor.RED + Language.MISC_ONE_USE.toString());
+        lore.add(ChatColor.GRAY + "+1 " + Language.ARG_PLOT);
         meta.setLore(lore);
-        if (Version.hasModelData()) meta.setCustomModelData(11);
+        if (Version.hasModelData()) meta.setCustomModelData(customModelData);
         meta.addItemFlags(ItemFlag.HIDE_ENCHANTS);
         meta.addEnchant(Enchantment.DURABILITY, 1, true);
         plotDeed.setItemMeta(meta);
         return plotDeed;
     }
 
-
-    public static ItemStack build() {
+    public ItemStack getItemStack() {
         return itemStack.clone();
     }
 
     public static boolean isPlotDeed(ItemStack itemStack) {
-        if (ItemUtil.hasLoreID(itemStack)) {
-            return plotDeedLoreID.equals(ItemUtil.getLoreID(itemStack));
+        if (hasLoreId(itemStack)) {
+            return plotDeedLoreID.equals(getLoreId(itemStack));
         } else {
             return false;
         }
     }
+
+    public static boolean hasLoreId(ItemStack itemStack) {
+        return (itemStack != null
+                && itemStack.hasItemMeta()
+                && itemStack.getItemMeta().hasLore()
+                && itemStack.getItemMeta().getLore().size() > 0
+                && itemStack.getItemMeta().getLore().get(0) != null
+        );
+    }
+
+    public static String getLoreId(ItemStack itemStack) {
+        return itemStack.getItemMeta().getLore().get(0);
+    }
+
 
 }